src/requirements.txt
certifi==2020.12.5
chainer==7.7.0
chainercv==0.13.1
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
fcn==6.4.17
filelock==3.0.12
gdown==3.12.2
idna==2.10
imageio==2.9.0
joblib==1.0.0
kaggle==1.5.10
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
pandas==1.1.5
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.1
scikit-image==0.18.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.12.8
torch==1.7.1
torchfcn==1.9.6
torchvision==0.8.2
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
